Change “Capital Allowances” to all lower case.
According to the this [1], “Capital Allowances”  should change to
“capital allowances”.

This commit shows the affected/changes to checksums and regression test
artefacts as a result of the content change.

I have reviewed the changes to the regression test artefacts and it shows
the difference as a result of the aforementioned change.

At this point existing unit, regression and integration tests passed, giving
assurance that the changes have had no (interfering) affect to other smart
answers.

[1] https://docs.google.com/document/d/1vyWhULg4cVGd4DJnhfBsbEtcOz077VeozWKSSdu0kSg/edit