Switch to use DateHelper
This switch allows the use of the ENV['RATES_QUERY_DATE']
environment variable.

This is helpful especially during fact-checking rates/content
dependent on future dates.

This will only be manually set and used in heroku.

Here, we still use Date.today especially when ENV['RATES_QUERY_DATE']
isn't set.. and for the purposes of fact-checking, we only manually
set this in heroku.. see more here [1]

[1] https://github.com/alphagov/smart-answers/blob/master/lib/smart_answer/date_helper.rb#L9